+++ src/contants.ts
@@ -1,7 +1,3 @@
 import { Logger } from "koishi"
 
-export const allows = {
-    systools: '3b10c8071d8c147b01e487b08d4073394b2f55f55071f16487d50438f015603c-$systools_send$'
-}
-
 export const logger = new Logger('milk-ikun')