+++ changes.md
@@ -1,3 +1,11 @@
+# V0.0.5
+1. add: 对 `Koishi Online` 环境的支持
+
+# V0.0.4
+1. fix: 安装并启用本插件后若环境中安装并启用了 `systools` 会导致本插件失效
+2. remove: 与 `systools` 的联动
+3. improve: `cultural reasons` => `cultural difference`
+
 # V0.0.3
 1. fix: `push.bat` 被上传
 2. add: `package.json` 的 `koishi` 字段
@@ -7,7 +15,6 @@
 # V0.0.2
 1. add: `package.json` github 仓库
 2. fix: `readme.md` 更新日志无法访问 的问题
-
 
 # V0.0.1
 1. add: 基础功能